backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

pymongo>=3.12.0
django-cors-headers>=4.2.0
channels>=4.0.0
daphne>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
dnspython>=2.3.0
